css/css-values/round-function.html

<!doctype html>
<title>round() function</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
</style>

<meta name=author content="Tab Atkins-Bittner">
<link rel=help href="https://drafts.csswg.org/css-values-4/#round-func">

<div></div>
<script>
const testEl = document.querySelector("div");
function testRound(testString, expectedString, {base="0px", msg}={}) {
    test(()=>{
        testEl.style.width = base;
        testEl.style.width = testString;
        const usedValue = getComputedStyle(testEl).width;
        assert_not_equals(usedValue, base);
        testEl.style.width = base;
        testEl.style.width = expectedString;
        const expectedValue = getComputedStyle(testEl).width;
        assert_not_equals(expectedValue, base)
        assert_equal(usedValue, expectedValue);
    }, msg || `${testString} should round to ${expectedString}`);
}

// No-op round should be same as nearest
testRound("round(23px, 10px)", "20px");
testRound("round(18px, 10px)", "20px");
testRound("round(15px, 10px)", "20px");
testRound("round(13px, 10px)", "10px");
testRound("round(-13px, 10px)", "-10px");
testRound("round(-18px, 10px)", "-20px");

// Test nearest
testRound("round(nearest, 23px, 10px)", "20px");
testRound("round(nearest, 18px, 10px)", "20px");
testRound("round(nearest, 15px, 10px)", "20px");
testRound("round(nearest, 13px, 10px)", "10px");
testRound("round(nearest, -13px, 10px)", "-10px");
testRound("round(nearest, -18px, 10px)", "-20px");

// Test down
testRound("round(down, 23px, 10px)", "20px");
testRound("round(down, 18px, 10px)", "10px");
testRound("round(down, 15px, 10px)", "10px");
testRound("round(down, 13px, 10px)", "10px");
testRound("round(down, -13px, 10px)", "-20px");
testRound("round(down, -18px, 10px)", "-20px");

// Test up
testRound("round(up, 23px, 10px)", "30px");
testRound("round(up, 18px, 10px)", "20px");
testRound("round(up, 15px, 10px)", "20px");
testRound("round(up, 13px, 10px)", "20px");
testRound("round(up, -13px, 10px)", "-10px");
testRound("round(up, -18px, 10px)", "-10px");

// Test to-zero
testRound("round(to-zero, 23px, 10px)", "20px");
testRound("round(to-zero, 18px, 10px)", "10px");
testRound("round(to-zero, 15px, 10px)", "10px");
testRound("round(to-zero, 13px, 10px)", "10px");
testRound("round(to-zero, -13px, 10px)", "-10px");
testRound("round(to-zero, -18px, 10px)", "-10px");

// Test a negative step
testRound("round(23px, -10px)", "20px");
testRound("round(18px, -10px)", "20px");
testRound("round(15px, -10px)", "20px");
testRound("round(13px, -10px)", "10px");
testRound("round(-13px, -10px)", "-10px");
testRound("round(-18px, -10px)", "-20px");

// Extreme cases:
function equalsPosInfinity(testString) {
    testRound(`calc(1px * ${testString})`, "calc(infinity * 1px)");
}
function equalsNegInfinity(testString) {
    testRound(`calc(1px * ${testString})`, "calc(-infinity * 1px)");
}
function equalsPosZero(testString) {
    testRound(`calc(1px / ${testString})`, "calc(infinity * 1px)");
}
function equalsNegZero(testString) {
    testRound(`calc(1px / ${testString})`, "calc(-infinity * 1px)");
}
function equalsNaN(testString) {
    // Make sure that it's NaN, not an infinity,
    // by making sure that it's the same value both pos and neg.
    testRound(`calc(1px * ${testString})`, "calc(NaN * 1px)");
    testRound(`calc(-1px * ${testString})`, "calc(NaN * 1px)");
}

// 0 step is NaN
equalsNaN("round(5, 0)");
// both infinite is NaN
equalsNaN("round(infinity, infinity)");

// infinite value with finite step is the same infinity
equalsPosInfinity("round(infinity, 5)");
equalsPosInfinity("round(infinity, -5)");
equalsNegInfinity("round(-infinity, 5)");
equalsNegInfinity("round(-infinity, -5)");

// Finite value with infinite step depends on rounding strategy.
// 'nearest' and 'to-zero': pos and +0 go to +0, neg and -0 go to -0
equalsPosZero("round(5, infinity)");
equalsPosZero("round(5, -infinity)");
equalsNegZero("round(-5, infinity)");
equalsNegZero("round(-5, -infinity)");
equalsPosZero("round(to-zero, 5, infinity)");
equalsPosZero("round(to-zero, 5, -infinity)");
equalsNegZero("round(to-zero, -5, infinity)");
equalsNegZero("round(to-zero, -5, -infinity)");
// 'up': pos goes to +inf, 0+ goes to 0+, else 0-
equalsPosInfinity("round(up, 1, infinity)");
equalsPosZero("round(up, 0, infinity)");
equalsNegZero("round(up, -1 * 0, infinity");
equalsNegZero("round(up, -1, infinity");
// 'down': neg goes to -inf, -0 goes to -0, else 0+
equalsNegInfinity("round(down, -1, infinity)");
equalsNegZero("round(down, -1 * 0, infinity)");
equalsPosZero("round(down, 0, infinity)");
equalsPosZero("round(down, 1, infinity)");
</script>
